+++ bitalarm-stack/src/main/java/com/dwarfeng/bitalarm/stack/service/AlarmSettingMaintainService.java
@@ -2,7 +2,7 @@ package com.dwarfeng.bitalarm.stack.service;
 
 import com.dwarfeng.bitalarm.stack.bean.entity.AlarmSetting;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
-import com.dwarfeng.subgrade.stack.service.CrudService;
+import com.dwarfeng.subgrade.stack.service.BatchCrudService;
 import com.dwarfeng.subgrade.stack.service.EntireLookupService;
 import com.dwarfeng.subgrade.stack.service.PresetLookupService;
 
@@ -12,7 +12,7 @@ import com.dwarfeng.subgrade.stack.service.PresetLookupService;
  * @author DwArFeng
  * @since 1.0.0
  */
-public interface AlarmSettingMaintainService extends CrudService<LongIdKey, AlarmSetting>,
+public interface AlarmSettingMaintainService extends BatchCrudService<LongIdKey, AlarmSetting>,
         EntireLookupService<AlarmSetting>, PresetLookupService<AlarmSetting> {
 
     String CHILD_FOR_POINT = "child_for_point";